refactor: make helper methods private
